+++ NEWS.md
@@ -1,5 +1,8 @@
 # ggplot2 2.2.1.9000
 
+* Custom objects can now be added using `+` if a `add_to_ggplot` method has been
+  defined for the class of the object (@thomasp85).
+
 * Fixed bug when setting strips to `element_blank()` (@thomasp85). 
 
 * Strips gain margins on all sides by default. This means that to fully justify